Edit grade

+++ backend.py
@@ -2,6 +2,7 @@ from flask import Flask, redirect, url_for, request, abort, render_template, fla
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin, login_manager, LoginManager, login_required, login_user, logout_user, current_user
 from flask_cors import CORS
+from random import randint
 
 app = Flask(__name__)
 
@@ -25,6 +26,14 @@ class User(db.Model, UserMixin):
 
     def check_password(self, password):
         return self.u_password == password
+
+    def is_teacher(self):
+        teacher = Teacher.query.filter_by(t_userId=self.u_userId).first()
+
+        if teacher != None:
+            return True
+
+        return False
 
 class Teacher(db.Model):
     t_teacherId = db.Column(db.Integer, primary_key=True)
@@ -95,7 +104,9 @@ def t_dashboard():
     # get student from student dashboard
     teacher = Teacher.query.filter_by(t_userId=current_user.u_userId).first()
 
-    return render_template('t_dashboard.html', name=teacher.t_name)
+    classesTaught = Class.query.filter_by(c_teacherId=teacher.t_teacherId).all()
+
+    return render_template('t_dashboard.html', name=teacher.t_name, classes=classesTaught)
 
 # Login requests
 @app.route('/login', methods=['GET', 'POST'])
@@ -133,7 +144,7 @@ def logout():
     logout_user()
     return {"redirect": url_for('index')}
 
-@app.route('/addClass/<classId>', methods=['GET', 'POST'])
+@app.route('/addClass/<int:classId>', methods=['GET', 'POST'])
 @login_required
 def addClass(classId):
     if request.method == 'POST':
@@ -148,7 +159,7 @@ def addClass(classId):
 
             #create enrollment for student
             allEnrollment = Enrollment.query.all()
-            e_id = len(allEnrollment) + 1
+            e_id = randint(0, 100_000)
             newEnrollment = Enrollment(e_id=e_id, e_classId=classId, e_studentId=student.s_studentId, e_grade=100.0)
             db.session.add(newEnrollment)
 
@@ -158,24 +169,57 @@ def addClass(classId):
         
     return '200'
 
-@app.route('/removeClass/<classId>', methods=['GET', 'POST'])
+@app.route('/removeClass/<int:classId>', methods=['GET', 'DELETE'])
 @login_required
 def removeClass(classId):
-    if request.method == 'POST':
+    if request.method == 'DELETE':
 
         course = Class.query.filter_by(c_classId=classId).first()
 
         student = Student.query.filter_by(s_userId=current_user.u_userId).first()
         
-        enrollment = Enrollment.query.filter_by(e_studentId=student.s_studentId).first()
+        enrollment = Enrollment.query.filter_by(e_classId=classId, e_studentId=student.s_studentId).first()
 
         # check if student is enrolled in class
-        if (enrollment != None):
+        if (enrollment != None and course.c_enrollmentNum != 0):
             db.session.delete(enrollment)
             course.c_enrollmentNum -= 1
             db.session.commit()
 
     return '200'
+
+@app.route('/lookupClass/<int:classId>', methods=['GET', 'POST'])
+@login_required
+def lookupClass(classId):
+    if request.method == 'GET':
+
+        #check for credentials
+        if(current_user.is_teacher()):
+
+            teacher = Teacher.query.filter_by(t_userId=current_user.u_userId).first()
+
+            students = Enrollment.query.filter_by(e_classId=classId).all()
+
+            class_name = Class.query.filter_by(c_classId=classId).first().c_courseName
+
+            return render_template('t_grades.html', name=teacher.t_name, students=students, class_name=class_name, Student=Student, classId=classId)
+
+    return '200'
+
+@app.route('/updateGrade', methods=['PUT'])
+@login_required
+def updateGrade():
+    print('updating...')
+    
+    if request.method == 'PUT':
+
+        enrollment = Enrollment.query.filter_by(e_classId=request.json['classId'], e_studentId=request.json['studentId']).first()
+
+        enrollment.e_grade = request.json['newGrade']
+
+        db.session.commit()
+
+    return '400'
 
 # Run app
 if __name__ == '__main__':
